refactor(settings): 拆出关于页贡献者数据模块与文件管理器页签三个分区
- 新增 settings/：appTabContributors（贡献者缓存/主题模式/接口拉取）、fileManager/FileManagerPreferencesSection、FileManagerConcurrencySection、FileManagerDownloadSection、fileManagerTabTypes
- AppTab.tsx 427 -> 301 行；FileManagerTab.tsx 553 -> 17 行（组合层）
