+++ frontend/src/components/settings/fileManager/fileManagerTabTypes.ts
@@ -1,0 +1,70 @@
+import React from 'react';
+
+export interface FileManagerTabProps {
+  fileManagerCompressedTransfer: boolean;
+  onToggleFileManagerCompressedTransfer: () => void;
+  fileManagerAutoOpenTransferQueue: boolean;
+  onToggleFileManagerAutoOpenTransferQueue: () => void;
+  fileManagerShowTabIcons: boolean;
+  onToggleFileManagerShowTabIcons: () => void;
+  fileManagerHideTabCloseButton: boolean;
+  onToggleFileManagerHideTabCloseButton: () => void;
+  fileManagerSharedPinnedTabs: boolean;
+  onToggleFileManagerSharedPinnedTabs: () => void;
+  fileManagerLayoutMode?: string;
+  onFileManagerLayoutModeChange?: (value: string) => void;
+  fileManagerDualPaneDragTransferEnabled: boolean;
+  onToggleFileManagerDualPaneDragTransferEnabled: () => void;
+  fileManagerDualPaneDragPromptOnDirectory: boolean;
+  onToggleFileManagerDualPaneDragPromptOnDirectory: () => void;
+  fileManagerDualPaneDragInvertModifier: boolean;
+  onToggleFileManagerDualPaneDragInvertModifier: () => void;
+  fileManagerChmodAutoApplyLastSettings: boolean;
+  onToggleFileManagerChmodAutoApplyLastSettings: () => void;
+  fileManagerDoubleClickUncompressArchive: boolean;
+  onToggleFileManagerDoubleClickUncompressArchive: () => void;
+  fileManagerSmartUncompressConflictStrategy?: string;
+  onFileManagerSmartUncompressConflictStrategyChange?: (value: string) => void;
+  fileManagerAutoRefreshDisabled?: boolean;
+  onToggleFileManagerAutoRefreshDisabled: () => void;
+  fileManagerMaxEditSizeMB?: number;
+  onFileManagerMaxEditSizeChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
+  fileManagerDefaultOpenMode?: string;
+  onFileManagerDefaultOpenModeChange?: (value: string) => void;
+  fileManagerPreferredExternalApp?: string;
+  onPickFileManagerPreferredExternalApp?: () => void;
+  onClearFileManagerPreferredExternalApp?: () => void;
+  fileManagerInitialPathMode: string;
+  onFileManagerInitialPathModeChange: (value: string) => void;
+  fileManagerNewTabPathMode: string;
+  onFileManagerNewTabPathModeChange: (value: string) => void;
+  fileManagerAskDownloadEveryTime: boolean;
+  onToggleFileManagerAskDownloadEveryTime: () => void;
+  fileManagerDownloadConflictStrategy: string;
+  onFileManagerDownloadConflictStrategyChange: (value: string) => void;
+  fileManagerDownloadConflictDiffBySize: boolean;
+  onToggleFileManagerDownloadConflictDiffBySize: () => void;
+  fileManagerDownloadConflictDiffByMtime: boolean;
+  onToggleFileManagerDownloadConflictDiffByMtime: () => void;
+  fileManagerDownloadRenameSuffixMode: string;
+  onFileManagerDownloadRenameSuffixModeChange: (value: string) => void;
+  fileManagerDownloadDefaultDir: string;
+  onFileManagerDownloadDefaultDirChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
+  fileManagerDownloadDefaultDirPreview: string;
+  fileManagerUploadChunkSizeKiB: number;
+  onFileManagerUploadChunkSizeKiBChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
+  fileManagerUploadMaxFiles: number;
+  onFileManagerUploadMaxFilesChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
+  fileManagerUploadMaxChunksPerFile: number;
+  onFileManagerUploadMaxChunksPerFileChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
+  fileManagerUploadGlobalInflightLimit: number;
+  onFileManagerUploadGlobalInflightLimitChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
+  transferMaxPacketKiB: number;
+  onTransferMaxPacketKiBChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
+  transferMaxRequestsPerFile: number;
+  onTransferMaxRequestsPerFileChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
+  transferConcurrentWrites: boolean;
+  onToggleTransferConcurrentWrites: () => void;
+  transferApplyToSharedClient: boolean;
+  onToggleTransferApplyToSharedClient: () => void;
+}
